+++ system/src/main/java/io/github/sds100/keymapper/system/files/FileAdapter.kt
@@ -9,6 +9,14 @@ interface FileAdapter {
     fun getPicturesFolder(): String
     fun openDownloadsFile(fileName: String, mimeType: String): KMResult<IFile>
 
+    /**
+     * Lists the files currently in the Downloads folder. If MANAGE_EXTERNAL_STORAGE
+     * is granted, this lists everything in the public Downloads directory. Otherwise it
+     * only lists files this app itself previously wrote there via [openDownloadsFile],
+     * which needs no permission.
+     */
+    fun getDownloads(): KMResult<List<IFile>>
+
     fun getPrivateFile(path: String): IFile
     fun getFile(parent: IFile, path: String): IFile
     fun getFileFromUri(uri: String): IFile