+++ system/src/main/java/io/github/sds100/keymapper/system/permissions/AndroidPermissionAdapter.kt
@@ -8,6 +8,7 @@ import android.content.Context
 import android.content.pm.IPackageManager
 import android.content.pm.PackageManager.PERMISSION_GRANTED
 import android.os.Build
+import android.os.Environment
 import android.os.PowerManager
 import android.os.Process
 import android.permission.IPermissionManager
@@ -345,6 +346,9 @@ class AndroidPermissionAdapter @Inject constructor(
             } else {
                 true
             }
+
+        Permission.MANAGE_EXTERNAL_STORAGE ->
+            Build.VERSION.SDK_INT >= Build.VERSION_CODES.R && Environment.isExternalStorageManager()
     }
 
     override fun isGrantedFlow(permission: Permission): Flow<Boolean> = channelFlow {