+++ system/src/main/java/io/github/sds100/keymapper/system/permissions/Permission.kt
@@ -19,4 +19,12 @@ enum class Permission {
     POST_NOTIFICATIONS,
     READ_LOGS,
     ACCESS_LOCAL_NETWORK,
+
+    /**
+     * Gives access to read all external files. Only requested
+     * on FOSS build because Google Play usually only permit it for
+     * file manager applications. This is needed on Android TV because
+     * it stubs the APIs for creating/reading documents.
+     */
+    MANAGE_EXTERNAL_STORAGE,
 }